Decouple texture bindings management from the texture cache
